requirements.txt
django==5.2.1
django-crispy-forms==2.4
pillow==11.2.1
python-dotenv==1.1.0
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2 